Replace the Decora integration with the newly created Decora BLE code
