+++ homeassistant/components/decora/const.py
@@ -1,0 +1,3 @@
+"""Constants for the Leviton Decora integration."""
+
+DOMAIN = "decora"
