+++ homeassistant/components/decora/light.py
@@ -1,151 +1,96 @@
-"""Support for Decora dimmers."""
+"""Decora BLE integration light platform."""
 from __future__ import annotations
 
-from collections.abc import Callable
-import copy
-from functools import wraps
-import logging
-import time
-from typing import TYPE_CHECKING, Any, Concatenate, ParamSpec, TypeVar
+from typing import Any, Optional
 
-from bluepy.btle import BTLEException
-import decora
-import voluptuous as vol
+from decora_bleak import DecoraBLEDevice, DecoraBLEDeviceState
 
-from homeassistant import util
-from homeassistant.components.light import (
-    ATTR_BRIGHTNESS,
-    PLATFORM_SCHEMA,
-    ColorMode,
-    LightEntity,
-)
-from homeassistant.const import CONF_API_KEY, CONF_DEVICES, CONF_NAME
-import homeassistant.helpers.config_validation as cv
+from homeassistant.components.light import ATTR_BRIGHTNESS, ColorMode, LightEntity
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers import device_registry as dr
+from homeassistant.helpers.device_registry import DeviceInfo
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-if TYPE_CHECKING:
-    from homeassistant.core import HomeAssistant
-    from homeassistant.helpers.entity_platform import AddEntitiesCallback
-    from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
+from .const import DOMAIN
+from .models import DecoraBLEData
 
 
-_DecoraLightT = TypeVar("_DecoraLightT", bound="DecoraLight")
-_R = TypeVar("_R")
-_P = ParamSpec("_P")
-
-_LOGGER = logging.getLogger(__name__)
-
-
-def _name_validator(config):
-    """Validate the name."""
-    config = copy.deepcopy(config)
-    for address, device_config in config[CONF_DEVICES].items():
-        if CONF_NAME not in device_config:
-            device_config[CONF_NAME] = util.slugify(address)
-
-    return config
-
-
-DEVICE_SCHEMA = vol.Schema(
-    {vol.Optional(CONF_NAME): cv.string, vol.Required(CONF_API_KEY): cv.string}
-)
-
-PLATFORM_SCHEMA = vol.Schema(
-    vol.All(
-        PLATFORM_SCHEMA.extend(
-            {vol.Optional(CONF_DEVICES, default={}): {cv.string: DEVICE_SCHEMA}}
-        ),
-        _name_validator,
-    )
-)
-
-
-def retry(
-    method: Callable[Concatenate[_DecoraLightT, _P], _R]
-) -> Callable[Concatenate[_DecoraLightT, _P], _R | None]:
-    """Retry bluetooth commands."""
-
-    @wraps(method)
-    def wrapper_retry(
-        device: _DecoraLightT, *args: _P.args, **kwargs: _P.kwargs
-    ) -> _R | None:
-        """Try send command and retry on error."""
-
-        initial = time.monotonic()
-        while True:
-            if time.monotonic() - initial >= 10:
-                return None
-            try:
-                return method(device, *args, **kwargs)
-            except (decora.decoraException, AttributeError, BTLEException):
-                _LOGGER.warning(
-                    "Decora connect error for device %s. Reconnecting",
-                    device.name,
-                )
-                # pylint: disable-next=protected-access
-                device._switch.connect()
-
-    return wrapper_retry
-
-
-def setup_platform(
+async def async_setup_entry(
     hass: HomeAssistant,
-    config: ConfigType,
-    add_entities: AddEntitiesCallback,
-    discovery_info: DiscoveryInfoType | None = None,
+    entry: ConfigEntry,
+    async_add_entities: AddEntitiesCallback,
 ) -> None:
-    """Set up an Decora switch."""
-    lights = []
-    for address, device_config in config[CONF_DEVICES].items():
-        device = {}
-        device["name"] = device_config[CONF_NAME]
-        device["key"] = device_config[CONF_API_KEY]
-        device["address"] = address
-        light = DecoraLight(device)
-        lights.append(light)
-
-    add_entities(lights)
+    """Set up the light platform for Decora BLE."""
+    data: DecoraBLEData = hass.data[DOMAIN][entry.entry_id]
+    async_add_entities([DecoraBLEEntity(data.device, data.address, entry.title)])
 
 
-class DecoraLight(LightEntity):
-    """Representation of an Decora light."""
+class DecoraBLEEntity(LightEntity):
+    """Representation of Decora BLE device."""
 
+    _attr_should_poll = False
+    _attr_has_entity_name = True
+    _attr_name = None
     _attr_color_mode = ColorMode.BRIGHTNESS
     _attr_supported_color_modes = {ColorMode.BRIGHTNESS}
 
-    def __init__(self, device: dict[str, Any]) -> None:
-        """Initialize the light."""
+    def __init__(self, device: DecoraBLEDevice, address: str, name: str) -> None:
+        """Initialize an Decora BLE light."""
+        super().__init__()
 
-        self._attr_name = device["name"]
-        self._attr_unique_id = device["address"]
-        self._key = device["key"]
-        self._switch = decora.decora(device["address"], self._key)
-        self._attr_brightness = 0
-        self._attr_is_on = False
+        self._device = device
+        self._attr_unique_id = address
+        self._address = address
 
-    @retry
-    def set_state(self, brightness: int) -> None:
-        """Set the state of this lamp to the provided brightness."""
-        self._switch.set_brightness(int(brightness / 2.55))
-        self._attr_brightness = brightness
+        summary = device.summary
+        if summary is not None:
+            self._attr_device_info = DeviceInfo(
+                name=name,
+                manufacturer=device.summary.manufacturer,
+                model=device.summary.model,
+                sw_version=device.summary.software_revision,
+                connections={(dr.CONNECTION_BLUETOOTH, self._address)},
+            )
+        else:
+            self._attr_device_info = DeviceInfo(
+                name=name,
+                connections={(dr.CONNECTION_BLUETOOTH, self._address)},
+            )
 
-    @retry
-    def turn_on(self, **kwargs: Any) -> None:
-        """Turn the specified or all lights on."""
-        brightness = kwargs.get(ATTR_BRIGHTNESS)
-        self._switch.on()
-        self._attr_is_on = True
+    @property
+    def available(self) -> bool:
+        """Determine if the entity is available."""
+        return bool(self._device.is_connected)
 
+    async def async_turn_on(self, **kwargs: Any) -> None:
+        """Instruct the light to turn on."""
+        brightness_level = self._to_device_brightness(kwargs.get(ATTR_BRIGHTNESS, None))
+        await self._device.turn_on(brightness_level)
+
+    async def async_turn_off(self, **kwargs: Any) -> None:
+        """Instruct the light to turn off."""
+        await self._device.turn_off()
+
+    async def async_added_to_hass(self) -> None:
+        """Register callbacks."""
+        self.async_on_remove(
+            self._device.register_state_callback(self._async_decora_state_changed)
+        )
+        return await super().async_added_to_hass()
+
+    @callback
+    def _async_decora_state_changed(self, state: DecoraBLEDeviceState) -> None:
+        """Handle state changed."""
+        self._attr_is_on = state.is_on
+        self._attr_brightness = self._from_device_brightness(state.brightness_level)
+
+        self.async_write_ha_state()
+
+    def _from_device_brightness(self, brightness_level: int) -> int:
+        return round((brightness_level / 100) * 255)
+
+    def _to_device_brightness(self, brightness: Optional[int]) -> Optional[int]:
         if brightness is not None:
-            self.set_state(brightness)
-
-    @retry
-    def turn_off(self, **kwargs: Any) -> None:
-        """Turn the specified or all lights off."""
-        self._switch.off()
-        self._attr_is_on = False
-
-    @retry
-    def update(self) -> None:
-        """Synchronise internal state with the actual light state."""
-        self._attr_brightness = self._switch.get_brightness() * 2.55
-        self._attr_is_on = self._switch.get_on()
+            return round((brightness / 255) * 100)
+        return None
