+++ homeassistant/components/decora/manifest.json
@@ -1,9 +1,16 @@
 {
   "domain": "decora",
   "name": "Leviton Decora",
-  "codeowners": [],
-  "documentation": "https://www.home-assistant.io/integrations/decora",
-  "iot_class": "local_polling",
-  "loggers": ["bluepy", "decora"],
-  "requirements": ["bluepy==1.3.0", "decora==0.6"]
+  "bluetooth": [
+    {
+      "service_uuid": "0000ff00-0000-1000-8000-00805f9b34fb"
+    }
+  ],
+  "codeowners": ["@ColinCampbell"],
+  "config_flow": true,
+  "dependencies": ["bluetooth_adapters"],
+  "documentation": "https://www.home-assistant.io/integrations/decora_ble",
+  "integration_type": "device",
+  "iot_class": "local_push",
+  "requirements": ["decora_bleak==0.2.2"]
 }
