+++ homeassistant/components/decora/strings.json
