--- homeassistant/components/decora_ble/const.py
@@ -1,3 +1,0 @@
-"""Constants for the Decora BLE integration."""
-
-DOMAIN = "decora_ble"
