--- homeassistant/components/decora_ble/manifest.json
@@ -1,16 +1,0 @@
-{
-  "domain": "decora_ble",
-  "name": "Decora BLE",
-  "bluetooth": [
-    {
-      "service_uuid": "0000ff00-0000-1000-8000-00805f9b34fb"
-    }
-  ],
-  "codeowners": ["@ColinCampbell"],
-  "config_flow": true,
-  "dependencies": ["bluetooth_adapters"],
-  "documentation": "https://www.home-assistant.io/integrations/decora_ble",
-  "integration_type": "device",
-  "iot_class": "local_push",
-  "requirements": ["decora_bleak==0.2.2"]
-}
