+++ mypy.ini
@@ -710,7 +710,7 @@ disallow_untyped_defs = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.decora_ble.*]
+[mypy-homeassistant.components.decora.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
